one Document type
